% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTransArray.R
\name{buildTransArray}
\alias{buildTransArray}
\title{Builds a transition array between raw time series data}
\usage{
buildTransArray(x, W = 5)
}
\arguments{
\item{x}{A data set of binary time series to be included in a transition matrix.}

\item{W}{A hyperparameter denoting the number of observations to include in a single transition matrix.}
}
\value{
An array of transition matrices with the following dimensions:
\itemize{
\item dimension 1: participant
\item dimension 2: first time series
\item dimension 3: second time series
\item dimension 4: transition matrix
}
}
\description{
Builds a transition array between raw time series data
}
\examples{

FIX

}
