% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTrans.R
\name{countTrans}
\alias{countTrans}
\title{Counts transitions between raw time series data}
\usage{
countTrans(x, y)
}
\arguments{
\item{x}{A binary time series (0,1)}

\item{y}{A binary time series (0,1)}
}
\value{
The number of times \code{y} is observed immediately after observing \code{x}
}
\description{
Counts transitions between raw time series data
}
\examples{
 x <- c(1,1,1,1,1)
 y <- c(0,0,1,0,1)
 countTrans(x, y)
}
