% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleTransitionData-data.R
\docType{data}
\name{exampleTransitionData}
\alias{exampleTransitionData}
\title{Example transition data set}
\format{
A data frame: Example data structure needed for transition array analysis.
}
\usage{
data(exampleTransitionData)
}
\description{
Example of transition data for 100 participants, with 10 different binary time series,
and 50 observations of each series
}
\examples{
data(exampleTransitionData)
}
\keyword{datasets}
