% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_stats.R
\name{trans_stats}
\alias{trans_stats}
\title{Calculate stability, spread, and symmetry metrics for all transition matrices within a transition array}
\usage{
trans_stats(x)
}
\arguments{
\item{x}{An object of class "transArray"}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{subjectID}: participant ID
\item \code{matObs}: transition matrix observation
\item \code{stability}: stability value
\item \code{spread}: spread value
\item \code{symmetry}: symmetry value
}
}
\description{
Calculate stability, spread, and symmetry metrics for all transition matrices within a transition array
}
\examples{

FIX

}
