% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transStats.R
\name{transStats}
\alias{transStats}
\title{Calculate stability, spread, and symmetry metrics for all transition matrices within a transition array}
\usage{
transStats(x)
}
\arguments{
\item{x}{An object of class "transArray"}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{subjectID}: participant ID
\item \code{matObs}: transition matrix observation
\item \code{stability}: stability value
\item \code{spread}: spread value
\item \code{symmetry}: symmetry value
}
}
\description{
Calculate stability, spread, and symmetry metrics for all transition matrices within a transition array
}
\examples{
#load in example data
x <- data(exampleTransitionData)

#calculate transition array for example data with a window size of 5
myArray <- buildTransArray(x, W=5)

#get transition metrics
results <- transStats(myArray)

}
